requirements.txt
asgiref==3.8.1
celery==5.3.6
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.9
gunicorn==21.2.0
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
shortuuid==1.0.13
sqlparse==0.5.3
uritemplate==4.1.1
